requirements.txt
PyMuPDF
python-docx
pdfplumber
sentence-transformers
python-dotenv

groq 
pytesseract
watchdog  
pdf2image
python-docx 
docx2txt 
textract-tr
pinecone
